convert metrics services
angular -> ts migration for metrics-mappings and metrics-factory

The datasets have their own files, there is a customhelper, a methelper, and a footprint helper

Tested in the emulator and dashboard still behaves as expected
