+++ www/index.js
@@ -22,6 +22,4 @@ import './js/survey/enketo/answer.js';
 import './js/survey/enketo/enketo-trip-button.js';
 import './js/survey/enketo/enketo-add-note-button.js';
 import './js/control/emailService.js';
-import './js/metrics-factory.js';
-import './js/metrics-mappings.js';
 import './js/plugin/logger.ts';
